+++ test/BasicTests/DynamicTests.cs
@@ -99,7 +99,7 @@ public class DynamicTests : TestMethods
                     print arr=dynamic(["this", "is", "an", "example"]) 
                     | project Result=array_reverse(arr)
                     | project A=Result[0]
-                    
+
                     """;
 
         var result = await LastLineOfResult(query);
@@ -126,5 +126,80 @@ public class DynamicTests : TestMethods
                            """);
     }
 
-    
+
+    [TestMethod]
+    public async Task ArrayIif1()
+    {
+        // query1
+        var query = """
+                    print condition=dynamic([true,false,true]), if_true=dynamic([1,2,3]), if_false=dynamic([4,5,6]) 
+                    | extend res= array_iif(condition, if_true, if_false)
+                    | extend res_str = strcat_array(res, ",")
+                    | project res_str
+                    """;
+        var expected = "1,5,3";
+        var res = await LastLineOfResult(query);
+
+        res.Should().Be(expected);
+    }
+
+    [TestMethod]
+    public async Task ArrayIif2()
+    {
+        // query2
+        var query2 = """
+                     print condition=dynamic([1,0,50]), if_true = dynamic(["yes"]), if_false=dynamic(["no"]) 
+                     | extend res= array_iif(condition, if_true, if_false)
+                     | extend res_str = strcat_array(res, ",")
+                     | project res_str
+                     """;
+        var expected2 = "yes,no,yes";
+        var res2 = await LastLineOfResult(query2);
+        res2.Should().Be(expected2);
+    }
+
+    [TestMethod]
+    public async Task ArrayIif3()
+    {
+        // //query 3
+        var query3 = """
+                     print condition=dynamic(["some string value", "datetime(01-01-2022)", null]), if_true=dynamic(["1"]), if_false=dynamic(["0"])
+                     | extend res= array_iif(condition, if_true, if_false)
+                     | extend res_str = strcat_array(res, ",")
+                     | project res_str
+                     """;
+        var expected3 = "null,null,null";
+        var res3 = await LastLineOfResult(query3);
+        res3.Should().Be(expected3);
+    }
+
+    [TestMethod]
+    public async Task ArrayIif4()
+    {
+        // query 4
+        var query4 = """
+                     print condition=dynamic([true,true,true]), if_true=dynamic([1,2]), if_false=dynamic([3,4]) 
+                     | extend res= array_iif(condition, if_true, if_false)
+                     | extend res_str = strcat_array(res, ",")
+                     | project res_str
+                     """;
+        var expected4 = "1,2,\"null\"";
+        var res4 = await LastLineOfResult(query4);
+        res4.Should().Be(expected4);
+    }
+
+    [TestMethod]
+    public async Task ArrayIff()
+    {
+        // query 4
+        var query4 = """
+                     print condition=dynamic([true,true,true]), if_true=dynamic([1,2]), if_false=dynamic([3,4]) 
+                     | extend res= array_iff(condition, if_true, if_false)
+                     | extend res_str = strcat_array(res, ",")
+                     | project res_str
+                     """;
+        var expected4 = "1,2,\"null\"";
+        var res4 = await LastLineOfResult(query4);
+        res4.Should().Be(expected4);
+    }
 }
