+++ test/BasicTests/SimpleFunctionTests.cs
@@ -1266,59 +1266,8 @@ print toscalar(letters | summarize mx=min(bitmap));";
         result.Should().Be(expected);
     }
 
-    [TestMethod]
-    public async Task array_iif()
-    {
-        // query1
-        var query = """
-                    print condition=dynamic([true,false,true]), if_true=dynamic([1,2,3]), if_false=dynamic([4,5,6]) 
-                    | extend res= array_iif(condition, if_true, if_false)
-                    | extend res_str = strcat_array(res, ",")
-                    | project res_str
-                    """;
-        var expected = "1,5,3";
-        var res = await LastLineOfResult(query);
-
-        res.Should().Be(expected);
-        
-        // query2
-        var query2 = """
-                    print condition=dynamic([1,0,50]), if_true = dynamic(["yes"]), if_false=dynamic(["no"]) 
-                    | extend res= array_iif(condition, if_true, if_false)
-                    | extend res_str = strcat_array(res, ",")
-                    | project res_str
-                    """;
-        var expected2 = "yes,no,yes";
-        var res2 = await LastLineOfResult(query2);
-        res2.Should().Be(expected2);
-
-        // //query 3
-        var query3 = """
-                    print condition=dynamic(["some string value", "datetime(01-01-2022)", null]), if_true=dynamic(["1"]), if_false=dynamic(["0"])
-                    | extend res= array_iif(condition, if_true, if_false)
-                    | extend res_str = strcat_array(res, ",")
-                    | project res_str
-                    """;
-        var expected3 = "null,null,null";
-        var res3 = await LastLineOfResult(query3);
-        res3.Should().Be(expected3);
-
-
-        // query 4
-        var query4 = """
-                    print condition=dynamic([true,true,true]), if_true=dynamic([1,2]), if_false=dynamic([3,4]) 
-                    | extend res= array_iif(condition, if_true, if_false)
-                    | extend res_str = strcat_array(res, ",")
-                    | project res_str
-                    """;
-        var expected4 = "1,2,\"null\"";
-        var res4 = await LastLineOfResult(query4);
-        res4.Should().Be(expected4);
-
-    }
-
+   
 
 
 }
 
-
